Make activity discussion embeddable
- Extract core into separate component
- Make heading optional
