+++ client/components/repository/common/Sidebar/Discussion/Embed.vue
@@ -1,0 +1,104 @@
+<template>
+  <div class="embedded-discussion">
+    <div :class="{ 'pb-7': !showHeading }">
+      <v-btn
+        v-if="commentsShownLimit < thread.length"
+        @click="showAll = !showAll"
+        text x-small
+        class="float-right mt-1">
+        Show {{ showAll ? 'less' : 'more' }}
+      </v-btn>
+    </div>
+    <div v-if="showHeading" class="header grey--text text--darken-3">
+      <v-icon color="grey darken-2" class="pr-1">mdi-forum-outline</v-icon>
+      Comments
+    </div>
+    <discussion-thread
+      v-if="thread.length"
+      :items="thread"
+      :user="user"
+      :min-displayed="commentsShownLimit"
+      :show-all="showAll"
+      class="mt-2" />
+    <div class="text-right">
+      <text-editor
+        ref="editor"
+        v-model="comment.content"
+        @change="post"
+        :placeholder="commentsCount ? 'Add a comment...' : 'Start the discussion...'" />
+      <v-btn @click="post" icon>
+        <v-icon>mdi-send</v-icon>
+      </v-btn>
+    </div>
+  </div>
+</template>
+
+<script>
+import { mapActions, mapGetters, mapState } from 'vuex';
+import DiscussionThread from './Thread';
+import orderBy from 'lodash/orderBy';
+import TextEditor from './TextEditor';
+
+const initCommentInput = () => ({ content: '' });
+
+export default {
+  name: 'embedded-discussion',
+  inheritAttrs: true,
+  props: {
+    activity: { type: Object, required: true },
+    showHeading: { type: Boolean, default: false }
+  },
+  data: () => ({ showAll: false, comment: initCommentInput() }),
+  computed: {
+    ...mapState({ user: state => state.auth.user }),
+    ...mapGetters('repository/comments', ['getActivityComments']),
+    comments: vm => vm.getActivityComments(vm.activity.id),
+    thread: vm => orderBy(vm.comments, ['createdAt'], ['asc']),
+    commentsCount: vm => vm.thread.length,
+    commentsShownLimit: vm => 5,
+    editor: vm => vm.$refs.editor.$el
+  },
+  methods: {
+    ...mapActions('repository/comments',
+      ['fetch', 'save', 'subscribe', 'unsubscribe']),
+    async post() {
+      if (!this.comment.content) return;
+      const payload = {
+        content: this.comment.content,
+        author: this.user,
+        activityId: this.activity.id,
+        createdAt: Date.now(),
+        updatedAt: Date.now()
+      };
+      this.comment = initCommentInput();
+      await this.save(payload);
+      // Keep editor inside viewport.
+      this.$nextTick(() => this.editor.scrollIntoView());
+    }
+  },
+  async created() {
+    await this.fetch(this.activity.id);
+    this.comment = initCommentInput();
+    this.subscribe();
+  },
+  beforeDestroy() {
+    this.unsubscribe();
+  },
+  components: {
+    DiscussionThread,
+    TextEditor
+  }
+};
+</script>
+
+<style lang="scss" scoped>
+.embedded-discussion {
+  font-family: Roboto, Arial, sans-serif;
+}
+
+.header {
+  margin: 0.875rem 0 1.625rem 0;
+  font-size: 1.125rem;
+  font-weight: 400;
+}
+</style>
